Extract `CoreData` helpers in their own file
